auto_gem-mcp_files/requirements.txt
google-genai
python-dotenv
pydantic